HMM/requirements.txt
Kivy==1.11.0
numpy==1.19.1
PyYAML==5.3.1
